refactor to support unittests
