+++ smart_scan/config.yml
